+++ transport/http/handler_test.go
@@ -78,7 +78,7 @@ func TestHandlerSucces(t *testing.T) {
 		gomock.Any(),
 	).Return(nil)
 
-	httpHandler := handler{Registry: registry, tracer: &opentracing.NoopTracer{}}
+	httpHandler := handler{registry: registry, tracer: &opentracing.NoopTracer{}}
 	req := &http.Request{
 		Method: "POST",
 		Header: headers,
@@ -131,7 +131,7 @@ func TestHandlerHeaders(t *testing.T) {
 			WithProcedure("hello"),
 		).Return(spec, nil)
 
-		httpHandler := handler{Registry: registry, tracer: &opentracing.NoopTracer{}}
+		httpHandler := handler{registry: registry, tracer: &opentracing.NoopTracer{}}
 
 		rpcHandler.EXPECT().Handle(
 			transporttest.NewContextMatcher(t,
@@ -258,7 +258,7 @@ func TestHandlerFailures(t *testing.T) {
 			).Return(spec, nil)
 		}
 
-		h := handler{Registry: reg, tracer: &opentracing.NoopTracer{}}
+		h := handler{registry: reg, tracer: &opentracing.NoopTracer{}}
 
 		rw := httptest.NewRecorder()
 		h.ServeHTTP(rw, tt.req)
@@ -309,7 +309,7 @@ func TestHandlerInternalFailure(t *testing.T) {
 		WithProcedure("hello"),
 	).Return(spec, nil)
 
-	httpHandler := handler{Registry: registry, tracer: &opentracing.NoopTracer{}}
+	httpHandler := handler{registry: registry, tracer: &opentracing.NoopTracer{}}
 	httpResponse := httptest.NewRecorder()
 	httpHandler.ServeHTTP(httpResponse, &request)